Continued cleaning up plugin to increase modularity, and reduce use of inline code.
-configuration.php
Added vis.min.js to vendor js file which is minified and served on page load. enables use of vis.js functions without going through a CDN.

-GraphController.php
changed php ajax request so that host  crts are not verified, enabling use of localhost adresses.

-display.phtml
removed all inline js code, moved to module.js

-module.js
created module.js file which is ran after page load, enables use of the already bundled jquery.

module.js now contains the ajax requests for the dependency and hosts json files from GraphController.php

-drawDependencies
added test on vis node creation to test if the current host is involved in the dependency tree, if not the host node is not created.

hasDependencies is now asserted if the node is identified if the host is identified as a parent. previously, no dependency information could exist for a parent, but the host could be referenced as the parent in other nodes.

-vis.min.js
added from vis website, is minified and served in icinga.min.js