Added testing  for validity of  entered login infromation through setup api CLI command.
-SetupCommand.php
added loop that repeats prompts for login information, makes an AJAX request  to server using provided login information. if basic ajax request is sucessful, login information is written to file. if there is an error, it is echoed back to user and then reprompted

the user is now also prompted for api address information, which is tested in the ajax request and stored in the file.

File is now stored in a .json instead of a .ini

-GraphController.php
now reads config information from config.json
corrected bug in $request_url