Added floating action button for editing/saving layout
-display.phtml
Added containers for zoom button, and save button.

-module.less
Added styling for FABs

-drawDependencies.js
cleaned up network options that are no longer needed.

added on click lisenters to edit FAB and save FAB. when edit is pressed, physics will be toggled on to allow for click and drag editing.

when save FAB is pressed, a JSON of current node positions is sent to back end where it is saved in a table.