When a host with dependencies is added or removed, the network is automatically regenerated.
-drawDependencies.js
after object generation, the hostObj(contains all hosts with dependency data) and the postionObj(contains stored host position data from last save) are compared, if the length is different, this signals that a host with dependency data has been added/removed.  positonObj is reset to be empty, signaling that the network should be redrawn and simulated.