A database is now used for settings, the plugin is now setup using a kickstart page that connects to a database.
-GraphController.php
removed all use of configuration file for getting api login information. All api information is pulled from the 'dependency' database, from a new table called "plugin settings".

Additionally, if the database does not exist or the login information is incorrect,  it echoes a code that causes the javascripti to redirect to the kickstart page for the plugin.

-pgsql.sql
added command to make new table to hold settings.

-kickstart.phtml
form for api/databse information  now uses post on submit, handled by jquery.

-getRequests.js
added new way to handle errors, which currently redirects (using a workarounds js redirect due to PHP's redirects  being solely in the header, which is already loaded when the plugin is started.) to the kickstart page assuming that most errors are caused by not linking the database, or having incorrect login information.

-kickstartManager.js
now uses jquery to POST to a new endpoint, storeSettings, which stores new settings in database.