Errors are now better handled on kickstart page.
-GraphController.php
When readding settings, table is truncated to prevent duplicates. as of right now, only one entry is needed.

When PHP performs a curl using loaded authentication information , errors are thrown  if the operation is unsuccessufull.

-kickstartManager.js
after entering information, An ajax request is made to put entries into database, if successful, anothjer ajax request is made to another endpoing to test the validity of entered information. if the server responds with anything other than a 200 code, the user is alerted to the details of the error whether the port, database, or authentication information was incorrect.