Improved Error Handling for every function that involves databases, and ajax requests.
-GraphController.php
On attempting to load the icinga database resource,  exceptions are caught depending if the resource doesnt exit or is empty. Correct Error headers and codes are sent to Icinga.

-drawDependencies.js
When recieiving an error from an ajax request,  alerts user to an error loading node positional data.

-getRequests.js
Each ajax request now supports alerting the user to a returned error.