Moved resource name back into a configuration file, enabling arbitrary naming for icinga resource linked to dependencies databse.
-GraphController.php
getresourcesAction() now reads name icinga resources name portion of the file, was previuously reading the database name, leading to incorrect behavior when populating the icinga resource list.

new function getResource() return the name of the selected resource, which is stored in config.ini, is called by every function that is interfacing with a database.

storesettingsAction() now stores settings in the settings table, and stores the name of the icinga database resource in a new file named "config.ini"

Every function now calls getResource() to get the name of the resource when attempting to use Icingas database management.

getRequests.js
Removed alert from every ajax request in the event of an error , was leading to 3 alerts for every backend error.

Now redirects to setup page in the event getResource echoes "setup" if the config file is not found, signifying first time use.

-init.sql
removed resource name column, now stored in config.ini