All hosts are now displayed, regardless if there is depdendency data.
-drawDependencies
When loading state information, if the host did not exist in the object created from dependencies, it meant that the host existed in icinga but had no dependency data, and was not graphed. Now the host is added and displayed.

-getRequests.js
moved welcome redirect to if there is no host loaded, instead of if there is no dependencies.