All user-customizable settings now implemented throughout module.
-configuration.php
changed default URL on dependency menu item, now redirects to 'home',  change was needed to implment the customization of the default graph type.

-GraphController.php
added homeAction() to display home.phtml

-home.phtml
Identical page to network.phtml, what type of network is displayed is decided by getRequests.

-getRequests.js
Added ajax requeste to load settings. after loading settings, checks if href is pointed at 'home'. if href is home, the value stored in the database for default network type is used to determine what network to render.

-drawDependencies.js
Now accepts a new object, 'settings'. Due to the varied nature of how vis js handles the various settings, the implementation of the user defined settings takes place in several differnt places throught the file.

to determine to display all hosts, or hosts with only dependencies,  a check is performed while iterating through the host JSON. If all hosts are to be displayed, hosts are added even if they are not found in object created by the dependencies.

For text size, node scaling, and displaying labels for hosts in a certain state,  settings are applied on a per node basis in drawNetwork().