Updated database to be more flexible when adding new settings,  instead of columns refering to settings names, settings are rows instead. this enables the adding of settings without database modifications.
-ModuleController.php
changed how values are inserted, and read from the table graph_settings. Additionally, if the table is not as expected it will be truncated and reformatted with default values. This eliminates the need for updates via a database commandline.

-init.sql
changed commands to reflect the creation of a graph_settings table formatted more simply, with only two columns, 'setting_name', 'setting_number'

-drawDependencies.js
changed how settings are checked during the drawing of the graph to reflect changes in the format of settings returned from the table 'graph_settings'

-getRequests.js
Now using for loop to parse passed settings which were formatted differently.

-settingsManager.js
changed how the settings object is initialized, and how values are parsed and stored to reflect changes in the format of the table 'graph_settings'