Added support for selecting a default director dependency template in the module settings page, as well as adding checks in drawDependencies.js that checks for a default tempelate before attempting to enter the edit dependency mode. The addition of this setting required the structure of how settings are stored into the db, and parsed on the front end.
-ModuleController.php
The graph_settings database now stores everything as a string, and has a new column 'setting_type' which denotes the format of the value stored when parsed on the efrontend, adding greater flexibility in adding settings.

the storegraphSettings() function now stores each setting as a string, in accordance with the new database schema.

getgraphsettingsAction() function now checks for correct number of columns as well as correct number of settings entries, if the table does not have the correct format, it automatically adds the correct column, eliminating the need to modify the table from commandline.

-init.sql
changed database to reflect hopefully (final) change to database schema.

-network.phtml
added new button to launch fullscreen

-settings.phtml
added new setting 'Default Dependency Template' which allows for the selection of a default dependency template.

-moudle.less
increased container width for FABs due to new button.

-drawDependencies.js
all settings no longer required to be parsed as an integer, as all settings formatting is now pre determined in getRequests()

added checks that ensure that a default dependency template is loaded.

added listeners for fullscreen button.

-fullscreenMode.js
removed tabs from fullscreen view.

-getRequests.js
now parses settings based on what 'type' they are.

-settingsManager.js
parses settings based on type, stores settings with a new attribute 'type'.

uses the director API to query dependency templates, and populate a dropdown.